Many Changes
Using FILES pointers is not doable in the long run because of
http://www.cplusplus.com/reference/cstdio/FILE/ “The content of a FILE
object is not meant to be accessed from outside the functions of the
<cstdio> and <cwchar> headers; In fact, portable programs shall only
use them in the form of pointers to identify streams, since for some
implementations, even the value of the pointer itself could be
significant to identify the stream (i.e., the pointer to a copy of a
FILE object could be interpreted differently than a pointer to the
original).”
Also to take care of labels it is necessary to know the address of the
instruction already in the IAS memory. So I started building a
compileProgram function.